src/requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
Fiona==1.9.1
geopandas==0.12.2
greenlet==2.0.2
idna==3.4
munch==2.5.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
psycopg2==2.9.5
pyproj==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
shapely==2.0.1
six==1.16.0
SQLAlchemy==2.0.6
typing_extensions==4.5.0
urllib3==1.26.15